Move filters to new model file
